fix(aws): do not create boto3 session if not necessary
In boto3, clients for specific AWS services can be created by first
creating a session, or directly from the module, in which case a default
session initialized by the module will be used.

It seems that when creating a session, that session will not always pick
up AWS credentials (empirically, it seems to pick up config from
~/.aws/, but not from the environment variables). The default session
seems to check all sources.

To allow maximum flexibility for how credentials may be configured, do
not create a Session unless custom parameters are specified (which, in
the current implementation, they never are), and rely on the default
when creating clients.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>

fix(aws): replace hard-coded permission boundary

Replace the hard-coded permission boundary ARN with the parameter that
was supposed to be used.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>

fix(aws): save RDS password after stack creation

Move saving the RDS password to cache to after stack creation. This way,
if the creation fails, it won't be saved. This is to handle the case
where creation fails because the stack already exists; saving the
new password would overwrite the old (correct) one.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>

fix(aws): update stack name in teardown

Update the name from "rds" to "support" this was renamed a while ago,
and this has fallen through the cracks.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>

feat: adding signed corim

Signed-off-by: Akhilesh Kr. Yadav <akhileshkr.yadav@Akhileshs-MacBook-Air.local>

feat: adding signed corim

Signed-off-by: Akhilesh Kr. Yadav <akhileshkr.yadav@Akhileshs-MacBook-Air.local>

feat: adding signed corim

Signed-off-by: Akhilesh Kr. Yadav <akhileshkr.yadav@Akhileshs-MacBook-Air.local>